Refactor common code out of libc wrappers.
Extract shared code for fetching shmfd from environment, unpacking it
and calling po_find() into a static function. Also clean up some of the
error handling logic so that we fall back to the behaviour of the
non-*at(2) system calls (e.g., using AT_FDCWD with the full path) when
we are unable to find or unpack the shared po_map. This also prevents
NULL pointer dereferences and assertion failures.