Set config.library in CMake-generated lit configuration.
The autoconfiguration path for lit was set up to find libpreopen.so, but
the CMake-generated path wasn't (so Jenkins couldn't run the tests). Fix
this problem using annoying CMake variables like
${CMAKE_SHARED_LIBRARY_SUFFIX} because we can't ask CMake, "what is the
path of this target we're building" any more (see CMake CMP0026) and
generator expressions can't be used in configure_file().